fix: EPMEDU-3821: App is crashing on the filling phone number step
  - Fix and improve visual transformation of georgian phone number (cursor position and selection)
